backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
firebase-admin==6.5.0
pydantic==2.7.1
python-dotenv==1.0.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
